+++ src/controllers/CustomersController.php
@@ -35,8 +35,8 @@ class CustomersController extends BaseCpController
      */
     public function init()
     {
-        $this->requirePermission('commerce-manageCustomers');
         parent::init();
+        $this->requirePermission('commerce-manageCustomers');
     }
 
     /**