+++ src/elements/traits/OrderElementTrait.php
@@ -9,6 +9,7 @@ namespace craft\commerce\elements\traits;
 
 use Craft;
 use craft\commerce\elements\actions\DeleteOrder;
+use craft\commerce\elements\actions\DownloadOrderPdf;
 use craft\commerce\elements\actions\UpdateOrderStatus;
 use craft\commerce\elements\db\OrderQuery;
 use craft\commerce\Plugin;
@@ -305,6 +306,11 @@ trait OrderElementTrait
 
         if (Craft::$app->getUser()->checkPermission('commerce-manageOrders')) {
             $elementService = Craft::$app->getElements();
+
+            if (Plugin::getInstance()->getPdfs()->getHasEnabledPdf()) {
+                $actions[] = DownloadOrderPdf::class;
+            }
+
             if (Craft::$app->getUser()->checkPermission('commerce-deleteOrders')) {
                 $deleteAction = $elementService->createAction(
                     [